Add user account to piksi_system_daemon, move uart config to ports_daemon
+ Move uart config the ports_daemon since it owns the serial devices

+ Remove async child and implement the handling of `upgrade_tool' with a
runit service.  This still needs to be completed by adding a facility to
send a SBP_MSG_COMMAND_RESP message after the upgrade completes.

+ Give piksi_system_daemon cap_net_admin so that it can
configure/interact with administrative bits of the system for network
interfaces without root.

+ Change the network configuration to shell out to a configure script so
that we can run just this portion as root.
