--- board/piksiv3/rootfs-overlay/etc/init.d/S82piksi_system_daemon
@@ -1,9 +1,0 @@
-#!/bin/sh
-
-name="piksi_system_daemon"
-cmd="piksi_system_daemon"
-dir="/"
-user=""
-
-source /etc/init.d/template_process.inc.sh
-
