+++ package/piksi_system_daemon/overlay/etc/init.d/S82piksi_system_daemon
@@ -1,0 +1,18 @@
+#!/bin/sh
+
+name="piksi_system_daemon"
+cmd="piksi_system_daemon"
+dir="/"
+user="piksi_sys"
+
+setup_permissions() {
+    # Setup runit dir dynamic services
+    mkdir -p /var/run/${name}/sv
+    chown -R ${user}:${user} /var/run/${name}
+
+    # Add the network admin capability
+    setcap cap_net_admin+ep /usr/bin/${name}
+}
+
+source /etc/init.d/template_process.inc.sh
+
