+++ package/piksi_system_daemon/overlay/etc/sudoers.d/piksi_system_daemon
@@ -1,0 +1,2 @@
+piksi_sys ALL=(ALL) NOPASSWD: /usr/bin/upgrade_tool --debug /data/upgrade.image_set.bin
+piksi_sys ALL=(ALL) NOPASSWD: /etc/init.d/update_eth0_config
