+++ package/piksi_system_daemon/overlay/etc/sv/piksi_system_daemon_runsvdir/run
@@ -1,0 +1,26 @@
+#!/bin/sh
+
+name=piksi_system_daemon
+user=piksi_sys
+
+service=${name}_runsvdir
+
+stderr=/var/run/${service}.stderr
+stdout=/var/run/${service}.stdout
+
+# enable job control
+set -m
+
+# kill -$pid will kill the whole process group
+trap 'kill -HUP -$pid && kill -TERM -$pid' HUP TERM
+
+rm -f $stderr $stdout
+mkfifo $stderr $stdout
+
+logger -t ${service} -p daemon.info <$stdout &
+logger -t ${service} -p daemon.err <$stderr &
+
+chpst -u ${user} runsvdir /var/run/${name}/sv 1>$stdout 2>$stderr &
+
+pid=$!
+wait $pid
