+++ package/ports_daemon/ports_daemon/src/Makefile
@@ -1,5 +1,5 @@
 TARGET=ports_daemon
-SOURCES=ports_daemon.c ports.c protocols.c whitelists.c
+SOURCES=ports_daemon.c ports.c protocols.c whitelists.c serial.c
 LIBS=-lczmq -lsbp -lpiksi -lzmq -ldl
 CFLAGS=-std=gnu11 -Wmissing-prototypes -Wimplicit -Wshadow -Wswitch-default -Wswitch-enum -Wundef -Wuninitialized -Wpointer-arith -Wstrict-prototypes -Wcast-align -Wformat=2 -Wimplicit-function-declaration -Wredundant-decls -Wformat-security -Wall -Wextra -Wno-strict-prototypes -Werror -ggdb3
