+++ package/ports_daemon/ports_daemon/src/serial.h
@@ -1,6 +1,6 @@
 /*
- * Copyright (C) 2017 Swift Navigation Inc.
- * Contact: Jacob McNamee <jacob@swiftnav.com>
+ * Copyright (C) 2018 Swift Navigation Inc.
+ * Contact: Swift Navigation <dev@swiftnav.com>
  *
  * This source is subject to the license found in the file 'LICENSE' which must
  * be be distributed together with this source. All other rights reserved.
@@ -10,15 +10,12 @@
  * WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
  */
 
-#ifndef SWIFTNAV_PORTS_H
-#define SWIFTNAV_PORTS_H
+#ifndef SWIFTNAV_SERIAL_H
+#define SWIFTNAV_SERIAL_H
 
-#include <stdint.h>
-#include <stdbool.h>
-#include <stdlib.h>
-#include <sys/wait.h>
 #include <libpiksi/settings.h>
 
-int ports_init(settings_ctx_t *settings_ctx);
+int serial_init(settings_ctx_t *settings_ctx);
 
-#endif /* SWIFTNAV_PORTS_H */
+#endif /* SWIFTNAV_SERIAL_H */
+
